Conflicts:
	Tools/ResourceEditor/Classes/SceneEditor/NodesPropertyControl.cpp

Former-commit-id: 426948619d0ba440bf2836ea690d9d524586cb82

+++ Projects/UnitTests/Teamcity/deployRunApp.sh
@@ -1,17 +1,35 @@
 # !/bin/bash
 echo "copy UnitTests.app"
-cp -Rf "../build/Release-iphoneos/UnitTests.app" "./"
+APP=UnitTests
+DIR_APP="../DerivedData/TemplateProjectiPhone/Build/Products/Release-iphoneos"
 
-echo "sign and create PerfomanceTest.ipa"
-sh floatsign.sh UnitTests.app 4L7VSNH4R3 UnitTests.ipa
+if [ ! -d $DIR_APP/$APP.app ]; then
+ echo "directory $DIR_APP/$APP.app doesn't exist"
+ exit 1
+fi
+cp -Rf $DIR_APP/$APP.app "./"
+
+echo "sign and create $APP.ipa"
+sh floatsign.sh $APP.app 4L7VSNH4R3 $APP.ipa
+sh floatsign.sh $APP.app 4L7VSNH4R2 $DIR_APP/$APP.ipa
+
+if [ ! -f $APP.ipa ]; then
+  echo "$APP.ipa wasn't created"
+  exit 1
+fi
+
 
 echo "deploy ipa on device"
-./transporter_chief.rb UnitTests.ipa
+./transporter_chief.rb $APP.ipa
+
+echo "get udid device" 
+UDID_device=`system_profiler SPUSBDataType | sed -n -e '/iPad/,/Serial/p' -e '/iPhone/,/Serial/p' -e '/iPod/,/Serial/p' | grep "Serial Number:" | awk -F ": " '{print $2}'`
+echo "udid is ${UDID_device}"
 
 echo "run app on device"
-instruments -w 3a80111143a4bf963e27ea94197a85d33039cec9 -t /Developer/Platforms/iPhoneOS.platform/Developer/Library/Instruments/PlugIns/AutomationInstrument.bundle/Contents/Resources/Automation.tracetemplate "UnitTests" -e UIASCRIPT testRun.js
+instruments -w ${UDID_device} -t /Developer/Platforms/iPhoneOS.platform/Developer/Library/Instruments/PlugIns/AutomationInstrument.bundle/Contents/Resources/Automation.tracetemplate "$APP" -e UIASCRIPT testRun.js
 
 #del temporary files
-rm -rf ./UnitTests.*
-rm -rf ./fruitstrap
+rm -rf ./$APP.*
+#rm -rf ./fruitstrap
 rm -rf ./instrumentscli*
